# config-ui-src/scca-rest/src/main/generated/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "This is a sample.",
    "version" : "v1",
    "title" : "Swagger Maven Plugin Sample",
    "termsOfService" : "http://www.github.com/kongchen/swagger-maven-plugin",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "${scca.rest.context-path}/encryptKey" : {
      "get" : {
        "tags" : [ "EncryptKey MGT（加密管理）" ],
        "summary" : "List EncryptKey / 需加密的Key清单",
        "description" : "",
        "operationId" : "findAll",
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespListEncryptKeyDto"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "EncryptKey MGT（加密管理）" ],
        "summary" : "Add EncryptKey / 新增需加密的Key",
        "description" : "",
        "operationId" : "createEncryptKey",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/EncryptKeyDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "EncryptKey MGT（加密管理）" ],
        "summary" : "Update EncryptKey / 更新需加密的Key",
        "description" : "",
        "operationId" : "updateEncryptKey",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/EncryptKeyDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "EncryptKey MGT（加密管理）" ],
        "summary" : "Delete EncryptKey / 删除需加密的Key",
        "description" : "",
        "operationId" : "deleteEncryptKey",
        "parameters" : [ {
          "name" : "encryptKeyId",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      }
    },
    "${scca.rest.context-path}/encryptKey/batch" : {
      "post" : {
        "tags" : [ "EncryptKey MGT（加密管理）" ],
        "summary" : "Add EncryptKeys / 批量新增需加密的Key",
        "description" : "",
        "operationId" : "createBatch",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/definitions/EncryptKeyDto"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "EncryptKey MGT（加密管理）" ],
        "summary" : "Delete EncryptKeys / 批量删除需加密的Key",
        "description" : "",
        "operationId" : "deleteBatch",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      }
    },
...
    "${scca.rest.context-path}/envParam" : {
      "get" : {
        "tags" : [ "EnvParam MGT（环境参数管理）" ],
        "summary" : "List EnvParam / 环境参数列表",
        "description" : "",
        "operationId" : "findEnvParamsByEnv",
        "parameters" : [ {
          "name" : "envId",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespListEnvParamDto"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "EnvParam MGT（环境参数管理）" ],
        "summary" : "Add EnvParam / 添加一个环境参数",
        "description" : "",
        "operationId" : "addEnvParam",
        "parameters" : [ {
          "name" : "envId",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/EnvParamDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "EnvParam MGT（环境参数管理）" ],
        "summary" : "Update EnvParam / 更新某个环境参数",
        "description" : "",
        "operationId" : "updateEnvParam",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/EnvParamDto"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "EnvParam MGT（环境参数管理）" ],
        "summary" : "Delete EnvParam / 删除某个环境参数",
        "description" : "",
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      }
    },
    "${scca.rest.context-path}/envParam/batch" : {
      "post" : {
        "tags" : [ "EnvParam MGT（环境参数管理）" ],
        "summary" : "Add EnvParams / 新增一批环境参数",
        "description" : "",
        "operationId" : "createBatch",
        "parameters" : [ {
          "name" : "envId",
          "in" : "query",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/definitions/EnvParamDto"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "EnvParam MGT（环境参数管理）" ],
        "summary" : "Delete EnvParams / 删除一批环境参数",
        "description" : "",
        "operationId" : "deleteBatch",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/WebRespString"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "EncryptKeyDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "ekey" : {
          "type" : "string"
        },
        "eKey" : {
          "type" : "string"
        }
      }
    },
...
    "EnvParamDto" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "pkey" : {
          "type" : "string"
        },
        "pvalue" : {
          "type" : "string"
        },
        "pKey" : {
          "type" : "string"
        },
        "pValue" : {
          "type" : "string"
        }
      }
    },
...
    "WebRespListEncryptKeyDto" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "data" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/EncryptKeyDto"
          }
        },
        "message" : {
          "type" : "string",
          "description" : "响应消息"
        },
        "timestamp" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "时间戳"
        }
      }
    },
...
    "WebRespListEnvParamDto" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "data" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/EnvParamDto"
          }
        },
        "message" : {
          "type" : "string",
          "description" : "响应消息"
        },
        "timestamp" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "时间戳"
        }
      }
    },
...
    "WebRespString" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "integer",
          "format" : "int32"
        },
        "data" : {
          "type" : "string"
        },
        "message" : {
          "type" : "string",
          "description" : "响应消息"
        },
        "timestamp" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "时间戳"
        }
      }
    },
...
  }
}